+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased
+
+* Fix `mmdf` collector to still write last collection metric during errors
+
 ## 1.5.0 / 2021-06-07
 
 * Add `gpfs_mmdiag_waiter_info` metric